LlamaBridge, LlamaModel, LlamaContext:
LlamaModel has been completely redesigned following the principles of Llama.cpp's llama-server.  It now stores the model context, sharing it among other instances of LlamaContext.  This allows having multiple chats run with the same model, dramatically reducing the memory requirements needed having multiple chats running... prior to this the library was loading the entire model multiple times.

ActionSystem:

ActionSystem has been updated so that Actions are lazy loaded when needed and that Actions have a dependency system in place with an ActionGroup.  This means you can do something like seen in "FunctionCallActionGroup": do a tone action, then run a FunctionCall and Chat action simultaneously, and then end with the SummaryAction once both FunctionCall and Chat action complete.   Furthermore, Actions can pass their results forward... such as the Chat action using the "Tone" results to drive how the response should be in tone to the user along with if the response should be short or long.  For the sample app this allows the Chat to end up being far less verbose unless the user asks for something in length, which is often difficult for a smaller model to do even with a good system prompt.

FunctionCall:

In this sample we have a demonstration of a Function system where there are 4 functions: control voice, change system prompt, remember name and "no-op" so it can decide not to do an action at all.

ObjectBox:

Messages, summaries, etc are saved in ObjectBox with the base framework for vector searches - but vector searching and string embedding has not been added in yet.
Overall:

Original design was on an iPad M series which was much more possible.  In the redesigned code it is now runs on an iPhone 15 Pro or 16+ series with decent performance.