Further enhancements to speed up functionality overall. 1. "message length" for short vs long moved to a custom TextClassifier to remove bottle-neck of LLM deciding that. 2. ActionRunner system updated to allow for LocalMLAction over just LLM. 3. Some bug fixes
Working on training a "FunctionPrediction" ml model to predict if LLM function should be called. Also, for actions that could be on/off switch related, some "functionality" can simply be mlmodel driven, reducing the reliance on using a large LLM to do all function calling.

Other change being considered is to allow thinking and chat models to be in their "queue", instead of just relying on Actions. This way function calling, summaries, etc can just happen independently when needed and chat runs when it needs.  The risk would be actions not quite matching current activity.